// src/types/ThemeColors.tsx
export type ThemeColors = {
    primary: string;
    text: string;
    textSecondary: string;

    background: string;
    backgroundElement: string;
    backgroundSelected: string;

    card: string;
    border: string;

    success: string;
    warning: string;
    danger: string;

    premium: string;
    premiumDark: string;
};